+++ PyWebRunner/runner.py
@@ -11,6 +11,7 @@ def main():
     parser.add_argument('-b', '--browser', help='Which browser to load. Defaults to Chrome.')
     parser.add_argument('--base-url', help='Base URL to use with goto command.')
     parser.add_argument('-t', '--timeout', help='Global wait timeout (in seconds). Defaults to 30.')
+    parser.add_argument('-do', '--default-offset', help='New default offset for scroll_to_element. (Default is 0)')
     parser.add_argument('--errors', dest='errors', action='store_true', help='Show errors.')
     parser.add_argument('--focus', dest='focus', action='store_true', help='Focus the browser on launch.')
     parser.add_argument('-v', '--verbose', dest='verbose', action='store_true', help='Verbose output of commands being executed.')
@@ -20,7 +21,8 @@ def main():
     errors = args.errors or False
     driver = args.browser or 'Chrome'
     timeout = args.timeout or 30
-    wt = WebTester(driver=driver, base_url=args.base_url, timeout=int(timeout))
+    default_offset = args.default_offset or 0
+    wt = WebTester(driver=driver, base_url=args.base_url, timeout=int(timeout), default_offset=default_offset)
     wt.start()
     if args.focus:
         wt.focus_browser()